refactor(form): split LoginFormActions (P12d)
Decompose the 478-LoC LoginFormActions god-module into 3 cohesive
sub-modules behind a 15-LoC re-export facade. No public-API change:
the six exports (IFillFromDiscoveryArgs, ISubmitResult, SubmitMethod
types; fillAllFields, fillAndSubmit, fillFromDiscovery functions)
keep their import paths byte-identical via the facade.

## Why

Phase 12d cluster: LoginFormActions.ts mixed three concerns —
classic mediator-driven fill, discovery-based fill via sealed
executor, and shared submit-bookkeeping. KG flagged it as the
last HIGH-tier complex file in the Login + Form cluster (file 8
of 8). Isolating the discovery path also gives the future §16.1
R-NO-CLEAR-LOG-CRED canary a clean lint surface on the fill
module without affecting the discovery branch.

## What

- Actions/ActionsTypes.ts: shared types (IFillAllResult,
  SubmitMethod, ISubmitResult, IFillAllArgs, IFillAndSubmitArgs,
  ISubmitPhaseResult, ISubmitPhaseArgs, IFillFromDiscoveryArgs)
  + helpers (normalizeSubmit, resolveSubmitMethod,
  resolveSubmitFromPhase, gateNoSubmitSignal, logFillCount,
  logSubmitResult).
- Actions/ActionsFill.ts: classic mediator path (runFieldReduce,
  fillAllFields, tryEnterSubmit, tryClickSubmit, runSubmitPhase,
  fillAndSubmit).
- Actions/ActionsDiscovery.ts: discovery-based path
  (logDiscoveryFill, fillOneDiscoveryField,
  validateDiscoveryCredentials, reduceDiscoveryFill,
  fillFieldsFromDiscovery, tryEnterFromDiscovery,
  mapClickRejection, tryClickSubmitFromDiscovery,
  submitViaDiscovery, runDiscoveryPrereqs, fillFromDiscovery).
- LoginFormActions.ts: 15-LoC facade re-exporting the public
  surface.

## Guideline compliance

- [x] coding-principle: SRP — fill / discovery / shared types split.
- [x] eslint-rules §3: no eslint-disable used.
- [x] eslint-rules §5: refactor commit does NOT touch eslint.config.
- [x] before-commit: never bypass hooks, never `git add .`, atomic.
- [x] commit §5: subject 47 chars; body wraps at <=72.
- [x] comments-in-code: BLUF JSDoc preserved on every export.
- [x] j-doc: @param/@return order; <p> for body paragraphs.

## Tracker(s)

- Phase 12d (Login + Form cluster decoupling — file 8 of 8, FINAL)
- PVTI_lAHOAPtF084BZbzwzgvauEI (GH Project #3)
- Master plan row 12 — v8.5.0 wave 2
